// _sdk/models/DeviceTokenRegistration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ChaChing Social Social API
 * FatFIRESocial is a social network for people who are interested in financial independence and early retirement. It is a place to share your journey, ask questions, and connect with others who are on the same path. Whether you are just starting out or have already achieved financial independence, FatFIRESocial is a community where you can find support, inspiration, and advice. Join us today and start building the life you want! ```
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: mabel.oza@chachingsocial.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DeviceTokenRegistration
 */
export interface DeviceTokenRegistration {
    /**
     * The unique identifier for the user.
     * @type {string}
     * @memberof DeviceTokenRegistration
     */
    userId: string;
    /**
     * The device token for the user.
     * @type {string}
     * @memberof DeviceTokenRegistration
     */
    deviceToken: string;
    /**
     * The platform of the user.
     * @type {string}
     * @memberof DeviceTokenRegistration
     */
    platform: DeviceTokenRegistrationPlatformEnum;
    /**
     * The version of the app.
     * @type {string}
     * @memberof DeviceTokenRegistration
     */
    appVersion?: string;
}


/**
 * @export
 */
export const DeviceTokenRegistrationPlatformEnum = {
    Ios: 'ios',
    Android: 'android',
    Web: 'web'
} as const;
export type DeviceTokenRegistrationPlatformEnum = typeof DeviceTokenRegistrationPlatformEnum[keyof typeof DeviceTokenRegistrationPlatformEnum];


/**
 * Check if a given object implements the DeviceTokenRegistration interface.
 */
export function instanceOfDeviceTokenRegistration(value: object): value is DeviceTokenRegistration {
    if (!('userId' in value) || value['userId'] === undefined) return false;
    if (!('deviceToken' in value) || value['deviceToken'] === undefined) return false;
    if (!('platform' in value) || value['platform'] === undefined) return false;
    return true;
}

export function DeviceTokenRegistrationFromJSON(json: any): DeviceTokenRegistration {
    return DeviceTokenRegistrationFromJSONTyped(json, false);
}

export function DeviceTokenRegistrationFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeviceTokenRegistration {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'],
        'deviceToken': json['deviceToken'],
        'platform': json['platform'],
        'appVersion': json['appVersion'] == null ? undefined : json['appVersion'],
    };
}

export function DeviceTokenRegistrationToJSON(json: any): DeviceTokenRegistration {
    return DeviceTokenRegistrationToJSONTyped(json, false);
}

export function DeviceTokenRegistrationToJSONTyped(value?: DeviceTokenRegistration | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'userId': value['userId'],
        'deviceToken': value['deviceToken'],
        'platform': value['platform'],
        'appVersion': value['appVersion'],
    };
}