// _sdk/models/DeviceTokenUnregistration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ChaChing Social Social API
 * FatFIRESocial is a social network for people who are interested in financial independence and early retirement. It is a place to share your journey, ask questions, and connect with others who are on the same path. Whether you are just starting out or have already achieved financial independence, FatFIRESocial is a community where you can find support, inspiration, and advice. Join us today and start building the life you want! ```
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: mabel.oza@chachingsocial.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DeviceTokenUnregistration
 */
export interface DeviceTokenUnregistration {
    /**
     * The unique identifier for the user.
     * @type {string}
     * @memberof DeviceTokenUnregistration
     */
    userId: string;
    /**
     * The device token for the user.
     * @type {string}
     * @memberof DeviceTokenUnregistration
     */
    deviceToken: string;
}

/**
 * Check if a given object implements the DeviceTokenUnregistration interface.
 */
export function instanceOfDeviceTokenUnregistration(value: object): value is DeviceTokenUnregistration {
    if (!('userId' in value) || value['userId'] === undefined) return false;
    if (!('deviceToken' in value) || value['deviceToken'] === undefined) return false;
    return true;
}

export function DeviceTokenUnregistrationFromJSON(json: any): DeviceTokenUnregistration {
    return DeviceTokenUnregistrationFromJSONTyped(json, false);
}

export function DeviceTokenUnregistrationFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeviceTokenUnregistration {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'],
        'deviceToken': json['deviceToken'],
    };
}

export function DeviceTokenUnregistrationToJSON(json: any): DeviceTokenUnregistration {
    return DeviceTokenUnregistrationToJSONTyped(json, false);
}

export function DeviceTokenUnregistrationToJSONTyped(value?: DeviceTokenUnregistration | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'userId': value['userId'],
        'deviceToken': value['deviceToken'],
    };
}